Return `None` from `CoalesceFormula`s with no components
`AggregationFormula`s return 0.0 when there are matching components.
For example, the battery `AggregationFormula` for a graph without
batteries would be 0.0.

This is not possible for `CoalesceFormula`s, so they were return an
`Err` in such cases until now.  This has been changed return `None`.

This is done by changing the design to repeatedly calling
`Expr::coalesce` for each item that needs to be added to a `coalesce`
expression.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>