Add consumer formula generator excluding phantom loads
Meters with successors can still have loads not represented in the
component graph.  These are called phantom loads.

Tracking phantom loads requires larger formulas and expects all meters
to be available.

This commit implements the consumer formula generator as it is
implemented in the SDK - exclude producer and battery powers from the
grid power, or when there is no grid meter, add all the non battery
and producer meters.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>